requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==41.0.7
kafka-python==2.0.2
celery==5.3.4
redis==5.0.1
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
django-redis==5.4.0
drf-spectacular==0.26.5
requests==2.31.0
